refactor: PointDto 리팩토링
