--- src/main/java/kr/hhplus/be/server/interfaces/dto/ChargePointRequestDto.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.interfaces.dto;
-
-import lombok.Data;
-
-@Data
-public class ChargePointRequestDto {
-    private Long amount;
-}
